docs/backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pytz
aiohttp
pydantic
sqlalchemy
